+++ src/cmds/help.js
@@ -1,12 +1,22 @@
-import { Command } from 'azk/cli/command';
+import { CliControllers } from 'cli-router';
+import { _, t } from 'azk';
+import { promiseResolve } from 'azk/utils/promises';
 
-class HelpCmd extends Command {
-  action(opts) {
-    this.parent.showUsage(opts.command);
-    return 0;
+class Help extends CliControllers.Help {
+  index(params, cli) {
+    var usage = super.index(params, cli);
+    usage = this.colorizeSections(params, usage);
+    this.ui.output(usage);
+    return promiseResolve(0);
+  }
+
+  colorizeSections(params, usage) {
+    _.map(this.sections, (section) => {
+      var regex = new RegExp(`^${section}:`, 'gmi');
+      usage = usage.replace(regex, t(`commands.help.${section}`));
+    });
+    return usage;
   }
 }
 
-export function init(cli) {
-  (new HelpCmd('help [command]', cli));
-}
+module.exports = Help;