+++ src/cmds/init.js
@@ -1,53 +1,50 @@
+import { CliController } from 'cli-router';
 import { _, config, fs, path, lazy_require, log } from 'azk';
-import { Command } from 'azk/cli/command';
+import { promiseResolve } from 'azk/utils/promises';
 
 var lazy = lazy_require({
   Generator: ['azk/generator'],
   example_system: ['azk/generator'],
 });
 
-class Cmd extends Command {
-  action(opts) {
-    if (opts.filename) {
+class Init extends CliController {
+  index(params) {
+    if (params.filename) {
       return this.showFilename();
     }
 
-    var manifest = config("manifest");
-    var cwd  = opts.path || this.cwd;
+    var generator = new lazy.Generator(this.ui);
+    var manifest  = config("manifest");
+    var cwd  = params.path || this.cwd;
     var file = path.join(cwd, manifest);
-    var generator = new lazy.Generator(this);
 
-    if (fs.existsSync(file) && !opts.force) {
-      this.fail(this.tKeyPath("already_exists"), manifest);
-      return 1;
+    if (fs.existsSync(file) && !params.force) {
+      this.ui.fail(this.ui.tKeyPath(this.name, "already_exists"), manifest);
+      return promiseResolve(1);
     }
 
     var systemsData = generator.findSystems(cwd);
     log.debug('generator.findSystems(\'%s\')', cwd);
 
     if (_.isEmpty(systemsData)) {
-      this.fail(this.tKeyPath("not_found"));
+      this.ui.fail(this.ui.tKeyPath(this.name, "not_found"));
       systemsData = { [lazy.example_system.name]: lazy.example_system };
     }
 
     generator.render({ systems: systemsData }, file);
-    this.ok(this.tKeyPath('generated'), manifest);
+    this.ui.ok(this.ui.tKeyPath(this.name, 'generated'), manifest);
 
     // Only show tips if is a git dir
     if (fs.existsSync(path.join(cwd, ".git"))) {
-      this.tOutput(this.tKeyPath('github'));
+      this.ui.tOutput(this.ui.tKeyPath(this.name, 'github'));
     }
 
-    return 0;
+    return promiseResolve(0);
   }
 
   showFilename() {
-    this.output(config('manifest'));
+    this.ui.output(config('manifest'));
   }
 }
 
-export function init(cli) {
-  return new Cmd('init [path]', cli)
-    .addOption(['--force', '-f'], { default: false })
-    .addOption(['--filename'], { default: false });
-}
+module.exports = Init;